Set feedPollDelay to 1ms
Optimize CosmosDB trigger config.

Henrik:
We have now also added "feedPollDelay": 1 in the cosmos database function.json file. The feedpolldelay has a default value of 5000ms, but since we only want focus on latency this was set to 1ms now. This worked nicely and we received an average of 59ms in latency.