Fix config reload for experiment plans
Changes in the of the `BENCHMARK_URL` are not updated when using an
experiment plan that changes triggers dynamically during runtime because
Dotenv does not override env variables by default.
This lead to the problem that the URL for the http trigger was used although
another trigger type was deployed. Hence, triggering failed and no appropriate traces
showed up in the log.

Using the `override=True` flag fixes this issue.